refector: matchView logic divide
